Lobby & General improvements:
- Added online column to server_data table to indicate when a server is online.
- If switching to offline server, switch server event will cancel.
- Portals.yml won't be loaded if file is default.
- Added method to open explore and building menu directly with the command /navigator explore and /navigator building
- Enabled lobby when starting plugin if server type is lobby.
- Added check if player is already in a portal as to not spam the inPortal event.
- Add /portals reload to reload portals from portals.yml
- Added teleport server event.

+++ src/main/java/me/bteuk/network/commands/Navigator.java
@@ -67,8 +67,10 @@ public class Navigator implements CommandExecutor {
 
     private void openExplore(NetworkUser u) {
 
-        u.mainGui.delete();
-        u.mainGui = null;
+        if (u.mainGui != null) {
+            u.mainGui.delete();
+            u.mainGui = null;
+        }
 
         u.mainGui = new ExploreGui(u);
         u.mainGui.open(u);
@@ -77,8 +79,10 @@ public class Navigator implements CommandExecutor {
 
     private void openBuilding(NetworkUser u) {
 
-        u.mainGui.delete();
-        u.mainGui = null;
+        if (u.mainGui != null) {
+            u.mainGui.delete();
+            u.mainGui = null;
+        }
 
         u.mainGui = new BuildGui(u);
         u.mainGui.open(u);
